navbar edit, homepage image edit move footer, change /profile
navbar edit, homepage image edit move footer, change /profile
