+++ frontend/src/components/footer.jsx
